requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.30